fix : 캐릭터 파츠 매터리얼도 변경 완료
