+++ Source/Rifa/RifaCharacterParts.cpp
@@ -12,15 +12,10 @@ ARifaCharacterParts::ARifaCharacterParts()
 {
  	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
 	PrimaryActorTick.bCanEverTick = false;
-	Root = CreateDefaultSubobject<USceneComponent>(TEXT("ROOT"));
-	Mesh = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("MESH"));
-	RootComponent = Root;
-	Mesh->SetupAttachment(Root);
-	static ConstructorHelpers::FObjectFinder<UStaticMesh> SW(TEXT("/Script/Engine.StaticMesh'/Game/RifaCharacters/Test/idletest_hair01.idletest_hair01'"));
-	if (SW.Succeeded())
-	{
-		Mesh->SetStaticMesh(SW.Object);
-	}
+	Mesh = CreateDefaultSubobject<USkeletalMeshComponent>(TEXT("MESH"));
+	Trigger = CreateDefaultSubobject<USphereComponent>(TEXT("TRIGGER"));
+	RootComponent = Trigger;
+	Mesh->SetupAttachment(RootComponent);
 	Mesh->SetCollisionProfileName(TEXT("NoCollision"));
 }
 
@@ -28,6 +23,20 @@ ARifaCharacterParts::ARifaCharacterParts()
 void ARifaCharacterParts::BeginPlay()
 {
 	Super::BeginPlay();
+	Trigger->OnComponentBeginOverlap.AddDynamic(this, &ARifaCharacterParts::OnCharacterOverlap);
+	Trigger->OnComponentEndOverlap.AddDynamic(this, &ARifaCharacterParts::EndCharacterOverlap);
+	Trigger->SetCollisionProfileName(TEXT("Trigger"));
+	if (IsValid(PickupTextClass))
+	{
+		PickupTextReference = Cast<UPickupText>(CreateWidget(GetWorld(), PickupTextClass));
+		if (IsValid(PickupTextReference))
+		{
+			PickupTextReference->PickupActor = Cast<AActor>(this);
+			PickupTextReference->PickupText = TEXT("Press E");
+			CharacterReference = Cast<ARifaCharacter>(UGameplayStatics::GetPlayerCharacter(GetWorld(), 0));
+			CharacterReference->PickupItem.AddDynamic(this, &ARifaCharacterParts::PickupCharacterParts);
+		}
+	}
 }
 
 // Called every frame
@@ -35,5 +44,39 @@ void ARifaCharacterParts::Tick(float DeltaTime)
 {
 	Super::Tick(DeltaTime);
 
+}
+
+void ARifaCharacterParts::PickupCharacterParts()
+{
+	if (GetActorEnableCollision() && IsInRange)
+	{
+		FName HairSocket(TEXT("hair_socket"));
+		PickupTextReference->RemoveFromParent();
+		SetActorHiddenInGame(true);
+		SetActorEnableCollision(false);
+		Trigger->RemoveFromRoot();
+		CharacterReference->CurrentHair = this;
+		CharacterReference->ChangeHairPart();
+	}
+
+}
+
+void ARifaCharacterParts::OnCharacterOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult)
+{
+	if (Cast<ARifaCharacter>(OtherActor))
+	{
+		PickupTextReference->AddToViewport();
+		IsInRange = true;
+	}
+
+}
+
+void ARifaCharacterParts::EndCharacterOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex)
+{
+	if (Cast<ARifaCharacter>(OtherActor))
+	{
+		PickupTextReference->RemoveFromParent();
+		IsInRange = false;
+	}
 }
